+++ src/parser/ast_common.h
@@ -37,7 +37,7 @@ typedef struct {
 
 typedef struct {
 	unsigned int minHops;
-	float maxHops;
+	unsigned int maxHops;
 } AST_LinkLength;
 
 typedef struct {
@@ -48,7 +48,7 @@ typedef struct {
 
 AST_NodeEntity* New_AST_NodeEntity(char *alias, char *label, Vector *properties);
 AST_LinkEntity* New_AST_LinkEntity(char *alias, char *label, Vector *properties, AST_LinkDirection dir, AST_LinkLength *length);
-AST_LinkLength* New_AST_LinkLength(unsigned int minHops, float maxHops);
+AST_LinkLength* New_AST_LinkLength(unsigned int minHops, unsigned int maxHops);
 AST_Variable* New_AST_Variable(const char *alias, const char *property);
 void Free_AST_GraphEntity(AST_GraphEntity *entity);
 void Free_AST_Variable(AST_Variable *v);
